Update ebuild to use cargo eclass
This change updates the ebuild to use the recently introduced cargo
eclass. This class is meant to be used for Rust programs using Cargo for
their build. In order to use this eclass we had to update the EAPI we
use to version 6. That in turn entailed switching from the git-2 eclass
to git-r3.
Note that we do not use all the features of the eclass. In particular,
we still install directly from the git repository as it includes all
Rust dependencies in tested versions.