Do not depend directly on dev-util/cargo
The cargo program can be provided by either the dev-util/cargo ebuild or
dev-lang/rust with the 'cargo' USE flag set. We should not force a
dependency on the former. All we really want to depend on is
virtual/cargo and said dependency is automatically added by the 'cargo'
eclass.
With this change we remove the explicit dependency and inherit from said
eclass.