Add ebuild for version 0.1.3
Most users are expected to necessarily require the most up-to-date
development snapshot of the program, but rather prefer what is expected
to be a potentially better tested version.
To this end, we will also have ebuilds for the individual versions of
the program. This change adds an ebuild for version 0.1.3 of the
program, as built directly from sources from crates.io.