Fix installation step
With a recent upgrade of cargo it appears that the originally used cargo
install command requires an additional argument to go through.
Unfortunately, the eclass does not supply this argument and so the build
fails.
To fix this problem, this change explicitly spells out the installation
command and adds the required argument.